pyproject.toml
[project]
name = "google-adk-streamlit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-cloud-secret-manager>=2.26.0",
    "google-genai==1.20.0",
    "streamlit==1.45.1",
]
